+++ services/company/controller/delete_company.ts
@@ -1,14 +1,17 @@
 import db from "../../../DB/db";
 
-async function delete_company(url:string) {
+async function delete_company(url:string,user:any) {
   try {
 
+    if(!user?.id)
+      return { status: 403, content: { message: "شما توانایی این کار را ندارید" } }
+    
     if(!url)
         return {status:400,content:'نام غرفه باید مشخص باشد'}
 
     let data:any=await db.execute(`DELETE FROM 
     company c
-    where c.english_name='${url}'`)
+    where c.url='${url}' AND for_user=${user.id}`)
 
     return {status:200,content:{message:'عملیات با موفقیت انجام شد'}}
   } catch (e) {
